Automatically updated pinned apps when running fix_shortcuts()
Make fix_shortcuts() more cautious in deleting shortcuts, only deleting
a shortcut from the labscript suite install directory if it has a matching
entry in the start menu and has the name of one of the apps in it.

This makes it safer for code to call fix_shortcuts() automatically.