Prevent race conditions in filewatcher stopping - modified info after shutdown is now guaranteeed to reflect file changes before stop() was called.
Introduce "clean modified info" that can be returned by a method, which
is not just what the user passed in, but also the initial info for files
that the user did not pass in a value for.

BLACS saving just the modified info made no sense. It needs to save the clean
modified into against which changes will be detected.

Compare file list and not just modificiation in order to determine whether
everything is in the initial state.

Only remove deleted files from the list of files to watch if they were not files
that were explicitly added (i.e. they are files within a tracked directory,
created after the watch behan). This  way, a deletion and restoration of a
watched file can trigger a 'original' event.